+++ src/extensions/technotype/technotypeext.cpp
@@ -86,7 +86,9 @@ TechnoTypeClassExtension::TechnoTypeClassExtension(const TechnoTypeClass *this_p
     MaxRandomSpawnOffset(0),
     IsDontScore(false),
     IsSpawned(false),
-    BuildTimeCost(0)
+    BuildTimeCost(0),
+    RequiredHouses(-1),
+    ForbiddenHouses(-1)
 {
     //if (this_ptr) EXT_DEBUG_TRACE("TechnoTypeClassExtension::TechnoTypeClassExtension - Name: %s (0x%08X)\n", Name(), (uintptr_t)(This()));
 }
@@ -321,6 +323,9 @@ bool TechnoTypeClassExtension::Read_INI(CCINIClass &ini)
     IsDontScore = ini.Get_Bool(ini_name, "DontScore", IsDontScore);
     IsSpawned = ini.Get_Bool(ini_name, "Spawned", IsSpawned);
     BuildTimeCost = ini.Get_Int(ini_name, "BuildTimeCost", BuildTimeCost);
+
+    RequiredHouses = ini.Get_Owners(ini_name, "RequiredHouses", RequiredHouses);
+    ForbiddenHouses = ini.Get_Owners(ini_name, "ForbiddenHouses", ForbiddenHouses);
 
     return true;
 }