+++ src/extensions/technotype/technotypeext.h
@@ -176,6 +176,18 @@ class TechnoTypeClassExtension : public ObjectTypeClassExtension
         bool IsSortCameoAsBaseDefense;
 
         /**
+         *  Bitfield of houses that can build this type.
+         *  If `RequiredHouses != -1`, only these houses can build it.
+         */
+        long RequiredHouses;
+
+        /**
+         *  Bitfield of houses that cannot build this type.
+         *  If `ForbiddenHouses != -1`, these houses cannot build it under any circumstances.
+         */
+        long ForbiddenHouses;
+
+        /**
          *  Description for the extended sidebar tooltip.
          */
         char Description[200];